Addtional/archive/autointerp_full_local/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "autointerp-full-local"
version = "0.1.1"
description = "Automated Interpretability with Local File Support"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["interpretability", "explainable-ai", "local-files"]
dependencies = [
    "torch",
    "datasets",
    "transformers",
    "orjson",
    "eai-sparsify",
    "safetensors",
    "simple_parsing",
    "jaxtyping",
    "fire",
    "blobfile",
    "bitsandbytes",
    "flask",
    "vllm",
    "aiofiles",
    "sentence_transformers",
    "anyio>=4.8.0",
    "faiss-cpu",
    "asyncer>=0.0.8",
    "beartype",
    "pyyaml"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pyright==1.1.378",
    "pytest-beartype", "pytest-asyncio", "setuptools", "pre-commit"
]
visualize = [
    "kaleido==0.2.1",
    "plotly>=5.0.0rc2",
    "pandas",
    "ipywidgets"
]

[tool.pyright]
include = ["autointerp_full_local*"]
reportPrivateImportUsage = false

[tool.setuptools.packages.find]
include = ["autointerp_full_local*"]

[tool.ruff]
# TODO: Clean up or remove experiments folder.
exclude = ["experiments", "**/prompts/*", "**/prompts.py","**/simulator.py"]
# Avoid automatically removing unused imports in __init__.py files.
# Such imports will be flagged with a dedicated message suggesting
# that the import is either added to the module's __all__ symbol
ignore-init-module-imports = true
# Ignore F821 (undefined name) errors, as they are common in type annotations.
ignore = ["F722", "F821"]
# Same as Black.
line-length = 88
# Enable pycodestyle (`E`), Pyflakes (`F`), and isort (`I`) codes
# See https://beta.ruff.rs/docs/rules/ for more possible rules
select = ["E", "F", "I"]

# Semantic release configuration removed - not needed for this package

[tool.pytest.ini_options]
beartype_packages = "autointerp_full_local"
